test(executions): extract state-store contracts and prove the in-memory fakes against them
Hoists the scenarios of AbstractExecutionStateStoreTest (5 tests) and
AbstractMultipleConditionStateStoreTest (22 tests) into annotation-free
contract superclasses (ExecutionStateStoreContract,
MultipleConditionStateStoreContract) with a store() seam. Zero scenarios
deleted or changed — they moved up one level verbatim:

- The legacy Abstract* classes keep their names, @MicronautTest,
  injected fields (incl. the protected executionStateStore the EE
  Elasticsearch subclass uses directly) and the flow() helper, so every
  JDBC (H2/Postgres/MySQL) and EE Elasticsearch subclass compiles and
  runs unchanged — verified: H2ExecutionStateStoreTest +
  H2MultipleConditionStateStoreTest run all 27 inherited scenarios
  green, and repository-elasticsearch-ee:compileTestJava passes.
- New InMemoryExecutionStateStoreTest and
  InMemoryMultipleConditionStateStoreTest (executor testkit) run the
  same 27 scenarios against the in-memory fakes in plain JUnit — the
  fakes behind every executor unit test are now provably faithful to
  the production store implementations.
